Move rind reward util under reward_score
